fix type check bug
